adminoobaa update alpha_tester, revert to stable angular 1.0.7
fixed in our expressjs error handler to return json on xhr (ajax)
requests, instead of the mistake that returned html.
however the xhr flag was not set by angular bug in version 1.1.5,
so because of this bug and the anoying added #/mydata bug, decided to
rever angular to stable version 1.0.7.
removed all ng-if usages because angular 1.0.7 doesn't have it.