Added folder creation test.
This requried that inode_create and inode_remove will be seperated
into two methods. The first communicating web req/res
The second an action method which is not aware of the req/res but
only of the required input e.g. the inode object to be created
and not a req containing the inode

test is running fine using:
./node_modules/grunt-contrib-nodeunit/node_modules/.bin/nodeunit star/routes/inode_api_test.js

but for some resason fails in:
 grunt nodeunit