issues #29 #41 - add SWM marker to prevent making changes on it
implemented the SWM checks in the client side to be able to give proper messages.
no checks for that on server side for now. its not critical.

added a loading icon to the share dialog until get_share_list returns.

resolved the TODO in common_api to return a function rather than use the
implicit bind() ways.

the star now also returns the inode owner in readdir (only when its not
the same as the requesting user). this is mostly a nice to have info,
because we don't yet support colaboration, but will be relevant once we
add colaborating on a folder.